// project/types/product.ts
export type Product = {
  id: string;
  name: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
  featured?: boolean;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
};